stock-bull/ml-engine/requirements.txt
# Core ML Libraries
scikit-learn>=1.4.1
xgboost==2.0.0
lightgbm==4.0.0
catboost==1.2.0

# Deep Learning (Optional)
tensorflow==2.13.0
keras==2.13.1

# Data Processing
pandas==2.1.0
numpy==1.24.3
scipy==1.11.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Model Interpretation
shap==0.42.1
lime==0.2.0.1

# Experiment Tracking
mlflow==2.6.0

# Utilities
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7